+++ app/Http/Controllers/RoleController.php
@@ -135,4 +135,11 @@ class RoleController extends ParentController
         return redirect()->route('roles.index')
             ->with('success', __('app.role_deleted'));
     }
+
+    public function permissions()
+    {
+        return view('roles.permissions', [
+            'permissions' => Config::get('auth.permissions')
+        ]);
+    }
 }